Compiles clean on Rocky 9.  Includes a couple of bugfixes related to NUL at end of sun_path field.

+++ sfio/src/lib/sfio/Stdio_b/snprintf.c
@@ -5,7 +5,7 @@
 */
 
 #if __STD_C
-int snprintf(char* s, int n,  const char* form, ...)
+int snprintf(char* s, long unsigned int n,  const char* form, ...)
 #else
 int snprintf(va_alist)
 va_dcl
